+++ lib/project_types/extension/features/runtimes/admin.rb
@@ -11,6 +11,7 @@ module Extension
           :port,
           :public_url,
           :renderer_version,
+          :resource_url,
           :shop,
           :uuid,
         ]